+++ Sources/StreamChatUI/ChatMessageList/Attachments/VideoAttachmentGalleryPreview.swift
@@ -85,14 +85,14 @@ open class VideoAttachmentGalleryPreview: _View, ThemeProvider {
         imageView.image = nil
         playButton.isVisible = false
 
-        if let url = content?.videoURL {
+        if let thumbnailURL = content?.thumbnailURL {
+            showPreview(using: thumbnailURL)
+        } else if let url = content?.videoURL {
             components.videoLoader.loadPreviewForVideo(at: url) { [weak self] in
                 self?.loadingIndicator.isHidden = true
-
                 switch $0 {
                 case let .success(preview):
-                    self?.imageView.image = preview
-                    self?.playButton.isVisible = self?.content?.uploadingState == nil
+                    self?.showPreview(using: preview)
                 case .failure:
                     break
                 }
@@ -101,6 +101,19 @@ open class VideoAttachmentGalleryPreview: _View, ThemeProvider {
 
         uploadingOverlay.content = content?.uploadingState
         uploadingOverlay.isVisible = uploadingOverlay.content != nil
+    }
+
+    private func showPreview(using thumbnailURL: URL) {
+        components.imageLoader.downloadImage(with: .init(url: thumbnailURL, options: ImageDownloadOptions())) { [weak self] result in
+            self?.loadingIndicator.isHidden = true
+            guard case let .success(image) = result else { return }
+            self?.showPreview(using: image)
+        }
+    }
+
+    private func showPreview(using thumbnail: UIImage) {
+        imageView.image = thumbnail
+        playButton.isVisible = content?.uploadingState == nil
     }
 
     /// A handler that is invoked when view is tapped.